requirements.txt
transformers==4.35.2
torch==2.1.1
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-cloud-aiplatform==1.36.4
pydantic==2.5.2
typing-extensions==4.8.0
starlette==0.27.0
requests==2.31.0